requirements.txt
requests==2.31.0
watchdog==4.0.0
python-dotenv==1.0.1
schedule==1.2.1